% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_seurat.R
\name{prepare_from_seurat}
\alias{prepare_from_seurat}
\title{Title}
\usage{
prepare_from_seurat(
  seuratObj,
  data.path,
  export.path,
  clusterNames = NULL,
  clusterGenes = NULL,
  top_5_genes = TRUE,
  multiSamplePattern = "_",
  verbose = FALSE
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{data.path}{path for 10x output directory}

\item{export.path}{path for output export}

\item{clusterNames}{names of the clusters}

\item{clusterGenes}{genes related to the clusters. This should be a list of arrays.}

\item{top_5_genes}{whether to be shown top fir genes at the spot on mouseover}

\item{multiSamplePattern}{the patterns to separate the multiple samples}

\item{verbose}{whether to print the process on colsole}
}
\value{

}
\description{
Title
}
