\name{XlibReplot}
\alias{XlibReplot}
\alias{getalst}
\alias{is.plot.window}
\alias{keepanc}
\alias{usenew}
\alias{setCallBack}
\alias{orig.zoom}
\alias{multipancPoint}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Zoom package utility functions
}
\description{
Different functions, should not be used by end users, definitions may change in future versions.
}
\usage{
XlibReplot(rp = NULL)
getalst(tmp=recordPlot()[[1]])
is.plot.window(alst)
keepanc(ancien,fact,new,point=NULL)
usenew(ancien,fact,new,point=NULL)
setCallBack(..., xlim = NULL, ylim = NULL, xaxs = "r", yaxs = "r")
orig.zoom(orig)
multipancPoint(ancien,fact=1,new,point=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rp}{
	Saved plot as generated by recordPlot().
}
  \item{tmp}{
	Usable part (equivalent to rp[[1]]) of a saved plot as generated by recordPlot().
}
  \item{alst}{
	Arguments part of an item in an recorded plot object.
}
  \item{ancien}{
	ancien set of coordinates
}
  \item{fact}{
	factor of magnification
}
  \item{new}{
	new set of coordinates
}
  \item{point}{
	coordinates of the pointer 
}
  \item{xlim}{
	x limits of the plot
}
  \item{ylim}{
	y limits of the plot
}
  \item{xaxs}{
	style of the x axis see xaxs in par() help
}
  \item{yaxs}{
	style of the y axis see yaxs in par() help
}
  \item{orig}{
	like rp, orig.zoom is at this point just a convenience wrapper for replayPlot()
}
  \item{...}{
	Possible addicional arguments to zoomplot.zoom
}
}
\details{
This functions are more or less tightly connected to R internals functions that are subject to a lot of changes and should not be relied on by a end user. 
In particular is.plot.windown() is responsible for the delicate task of identifying in a recording of a plot to identify the part actually responsible for plotting the graphic window. This single thing has changed three times between 2010 and 2013.

setCallBack() sets up the call handler. It should probably never be called alone.

keepanc, usenew and multipancPoint are three functions possibly called by zoomplot.zoom() in an interchangeable way, none of them make use of all the arguments. 
keepanc() and usenew() return ancien or new according to their names.
multipancPoint() allows to transform "ancien" into new coordinates, after magnification by "fact" (ignores "new") but keeping point invariant.

}
\value{
Can't be relied upon at this point.
}

\author{
Corentin M. Barbu
}

\seealso{
zm, zoomplot.zoom
}
\examples{
multipancPoint(c(0,1),fact=2,NULL,point=0.1)
}

