% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoomplot.zoom}
\alias{zoomplot.zoom}
\title{Central low level function of the zoom package.}
\usage{
zoomplot.zoom(
  xlim = NULL,
  ylim = NULL,
  fact = NULL,
  moveX = NULL,
  moveY = NULL,
  rp = NULL,
  x = NULL,
  y = NULL,
  xlimfn = NULL,
  ylimfn = NULL,
  ...
)
}
\arguments{
\item{xlim}{A vector with min and max x}

\item{ylim}{A vector with min and max y}

\item{fact}{A scalar giving the magnification factor (>1 brings you closer)}

\item{moveX}{Expected shift on X axis.}

\item{moveY}{Expected shift on Y axis.
corresponding warnings in ?recordPlot.}

\item{rp}{A previously recorded plot with recordPlot(). With all the}

\item{x}{x of a fix point when rescaling, by default the center.}

\item{y}{y of a fix point when rescaling, by default the center.}

\item{xlimfn}{a function using x, y and/or fact to generate new x lim if NULL and xlim/ylim not given will use multipancPoint}

\item{ylimfn}{a function using x, y and/or fact to generate new y lim, if NULL will use xlimfn}

\item{\dots}{Additional parameters not implemented, just in case.}
}
\value{
Not guaranted for now.
}
\description{
This function allows to replot the current or a saved plot with specific
boundaries, magnification factor and possibly arround a user defined x/y.
}
\details{
This function is not necessarily easy to use by hand. It is designed to work
well when called from higher level functions. End user should always use
zm().
}
\note{
This function is the heart of the zoom package and the one that can be.
affected by R version changes.
It is inspired by the zoomplot function in TeachingDemos package
}
\examples{

plot(rnorm(1000),rnorm(1000))
zoomplot.zoom(fact=2,x=0,y=0)

}
\seealso{
zm, in.zoom
}
\author{
Corentin M. Barbu
}
\keyword{plot}
\keyword{zoom}
