\name{in.zoom}
\alias{in.zoom}
\alias{out.zoom}
\alias{inout.zoom}
\alias{sq.zoom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Direct access to zoom functionalities.
}
\description{
These are not intended for direct end user use. They may be used in scripts.
}
\usage{
in.zoom(...)
out.zoom(...)
inout.zoom(...)
sq.zoom(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
	Extra arguments to zoomplot.zoom.
}
}
\value{
No guaranted output value
}

\author{
Corentin M. Barbu
}
\seealso{
zm(), session.zoom().
}

