% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{in.zoom}
\alias{in.zoom}
\alias{out.zoom}
\alias{set.zoom}
\alias{sq.zoom}
\alias{move.to.click.zoom}
\alias{inout.zoom}
\title{Direct access to zoom functionalities.}
\usage{
in.zoom(...)

move.to.click.zoom(...)

inout.zoom(...)

out.zoom(...)

set.zoom(...)

sq.zoom(...)
}
\arguments{
\item{\dots}{Extra arguments to zoomplot.zoom.}
}
\description{
Direct selection of a zoom method of the "session" type.
Possibly of use in scripts?
}
\note{
Each function starts a different interactive sequence
 	\itemize{
	 	\item{inout.zoom(): }{left click within bounds zooms in, outside bounds zoom out}
	 	\item{move.to.click.zoom(): }{center plot arround left click}
	 	\item{in.zoom(): }{each left click zooms in}
	 	\item{out.zoom(): }{each left click zooms out}
	 	\item{set.zoom(): }{ask for a magnification factor}
		\item{sq.zoom(): }{allow to click on the two corners of the desired region to zoom on}
}
}
\seealso{
zm(), session.zoom().
}
\author{
Corentin M. Barbu
}
