% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom-package.R
\docType{package}
\name{zoom-package}
\alias{zoom-package}
\alias{zoom}
\title{The zoom Package}
\description{
A spatial data visualization tool.
}
\details{
\tabular{ll}{
Package: \tab zoom\cr
Type: \tab Package\cr
Version: \tab 2.0.6\cr
Date: \tab 2014-03-18\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://github.com/cbarbu/R-package-zoom\cr
}

zm(), enters an interactive session to zoom and navigate the active plot. The development version, as well as binary releases can be found at https://github.com/cbarbu/R-package-zoom
}
\author{
Corentin M Barbu \email{corentin.barbu@gmail.com}, with contributions from Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{package}
